backend/demo_test.py
#!/usr/bin/env python3
"""
Demo Test Script for Futures Link Assistant
Tests all key functionality before demo
"""

import requests
import json
import time

BASE_URL = "http://localhost:5001"

def test_endpoint(endpoint, method="GET", data=None):
    """Test an endpoint and return the result"""
    try:
        if method == "GET":
            response = requests.get(f"{BASE_URL}{endpoint}")
        elif method == "POST":
            response = requests.post(f"{BASE_URL}{endpoint}", json=data)
        
        print(f"✅ {method} {endpoint}: {response.status_code}")
        if response.status_code == 200:
            try:
                return response.json()
            except:
                return {"message": "Response received"}
        else:
            print(f"   ❌ Error: {response.text}")
            return None
    except Exception as e:
        print(f"❌ {method} {endpoint}: {str(e)}")
        return None

def main():
    print("🚀 Futures Link Assistant - Demo Test")
    print("=" * 50)
    
    # Test basic endpoints
    print("\n1. Testing Basic Endpoints:")
    test_endpoint("/api/health")
    test_endpoint("/api/demo_status")
    test_endpoint("/api/campuses")
    
    # Test greeting audio
    print("\n2. Testing Greeting Audio:")
    greeting_result = test_endpoint("/api/greeting_audio")
    if greeting_result:
        print("   ✅ Greeting audio endpoint working")
    
    # Test voice processing
    print("\n3. Testing Voice Processing:")
    test_data = {
        "text": "South campus had 150 people, 8 new visitors, 3 salvations",
        "campus": "south"
    }
    voice_result = test_endpoint("/api/process_voice", "POST", test_data)
    if voice_result:
        print("   ✅ Voice processing working")
        if "stats" in voice_result:
            print(f"   📊 Extracted stats: {voice_result['stats']}")
    
    # Test query functionality
    print("\n4. Testing Query System:")
    query_data = {
        "question": "How many new people has south campus had this year?"
    }
    query_result = test_endpoint("/api/query", "POST", query_data)
    if query_result:
        print("   ✅ Query system working")
        if "answer" in query_result:
            print(f"   💬 Answer: {query_result['answer']}")
    
    print("\n" + "=" * 50)
    print("🎉 Demo Test Complete!")
    print("\nDemo Checklist:")
    print("✅ Backend running")
    print("✅ Health check working")
    print("✅ Greeting audio ready")
    print("✅ Voice processing working")
    print("✅ Query system working")
    print("✅ Campus detection working")
    print("\nReady for demo! 🚀")

if __name__ == "__main__":
    main() 