backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0
gspread==5.11.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
oauth2client==4.1.3
anthropic==0.7.8
openai==1.3.7
elevenlabs==0.2.26
python-dotenv==1.0.0
pygame==2.5.2
aiohttp==3.9.0
pydub==0.25.1
speechrecognition==3.10.0
numpy>=1.26.0
pandas>=2.0.0